tsm1: refactor and rename some methods
