+++ tsdb/tsm1/reader_index_test.go
@@ -108,49 +108,49 @@ func TestIndirectIndex_DeleteRange(t *testing.T) {
 
 	check(t, ind.Contains([]byte("mem")), true)
 	check(t, ind.Contains([]byte("cpu1")), true)
-	check(t, ind.ContainsValue([]byte("cpu1"), 4), true)
-	check(t, ind.ContainsValue([]byte("cpu1"), 5), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 10), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 15), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 16), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 4), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 5), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 10), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 15), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 16), true)
 	check(t, ind.Contains([]byte("cpu2")), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 4), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 5), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 10), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 15), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 16), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 4), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 5), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 10), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 15), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 16), true)
 
 	ind.DeleteRange([][]byte{[]byte("cpu1"), []byte("cpu2")}, 0, 5)
 
 	check(t, ind.Contains([]byte("mem")), true)
 	check(t, ind.Contains([]byte("cpu1")), true)
-	check(t, ind.ContainsValue([]byte("cpu1"), 4), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 5), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 10), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 15), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 16), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 4), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 5), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 10), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 15), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 16), true)
 	check(t, ind.Contains([]byte("cpu2")), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 4), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 5), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 10), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 15), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 16), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 4), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 5), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 10), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 15), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 16), true)
 
 	ind.DeleteRange([][]byte{[]byte("cpu1"), []byte("cpu2")}, 15, 20)
 
 	check(t, ind.Contains([]byte("mem")), true)
 	check(t, ind.Contains([]byte("cpu1")), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 4), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 5), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 10), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 15), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 16), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 4), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 5), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 10), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 15), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 16), false)
 	check(t, ind.Contains([]byte("cpu2")), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 4), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 5), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 10), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 15), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 16), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 4), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 5), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 10), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 15), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 16), false)
 }
 
 func TestIndirectIndex_DeletePrefix(t *testing.T) {
@@ -173,49 +173,49 @@ func TestIndirectIndex_DeletePrefix(t *testing.T) {
 
 	check(t, ind.Contains([]byte("mem")), true)
 	check(t, ind.Contains([]byte("cpu1")), true)
-	check(t, ind.ContainsValue([]byte("cpu1"), 4), true)
-	check(t, ind.ContainsValue([]byte("cpu1"), 5), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 10), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 15), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 16), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 4), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 5), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 10), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 15), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 16), true)
 	check(t, ind.Contains([]byte("cpu2")), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 4), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 5), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 10), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 15), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 16), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 4), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 5), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 10), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 15), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 16), true)
 
 	ind.DeletePrefix([]byte("cp"), 0, 5)
 
 	check(t, ind.Contains([]byte("mem")), true)
 	check(t, ind.Contains([]byte("cpu1")), true)
-	check(t, ind.ContainsValue([]byte("cpu1"), 4), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 5), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 10), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 15), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 16), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 4), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 5), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 10), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 15), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 16), true)
 	check(t, ind.Contains([]byte("cpu2")), true)
-	check(t, ind.ContainsValue([]byte("cpu2"), 4), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 5), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 10), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 15), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 16), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 4), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 5), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 10), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 15), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 16), true)
 
 	ind.DeletePrefix([]byte("cpu"), 15, 20)
 
 	check(t, ind.Contains([]byte("mem")), true)
 	check(t, ind.Contains([]byte("cpu1")), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 4), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 5), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 10), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 15), false)
-	check(t, ind.ContainsValue([]byte("cpu1"), 16), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 4), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 5), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 10), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 15), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu1"), 16), false)
 	check(t, ind.Contains([]byte("cpu2")), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 4), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 5), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 10), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 15), false)
-	check(t, ind.ContainsValue([]byte("cpu2"), 16), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 4), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 5), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 10), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 15), false)
+	check(t, ind.MaybeContainsValue([]byte("cpu2"), 16), false)
 }
 
 func TestIndirectIndex_DeletePrefix_NoMatch(t *testing.T) {
@@ -234,7 +234,7 @@ func TestIndirectIndex_DeletePrefix_NoMatch(t *testing.T) {
 	ind.DeletePrefix([]byte("d"), 5, 5)
 
 	check(t, ind.Contains([]byte("cpu")), true)
-	check(t, ind.ContainsValue([]byte("cpu"), 5), true)
+	check(t, ind.MaybeContainsValue([]byte("cpu"), 5), true)
 }
 
 //
