+++ tsdb/tsm1/writer.go
@@ -275,6 +275,43 @@ type directIndex struct {
 	indexEntries *indexEntries
 }
 
+type indexEntries struct {
+	Type    byte
+	entries []IndexEntry
+}
+
+func (a *indexEntries) Len() int      { return len(a.entries) }
+func (a *indexEntries) Swap(i, j int) { a.entries[i], a.entries[j] = a.entries[j], a.entries[i] }
+func (a *indexEntries) Less(i, j int) bool {
+	return a.entries[i].MinTime < a.entries[j].MinTime
+}
+
+func (a *indexEntries) MarshalBinary() ([]byte, error) {
+	buf := make([]byte, len(a.entries)*indexEntrySize)
+
+	for i, entry := range a.entries {
+		entry.AppendTo(buf[indexEntrySize*i:])
+	}
+
+	return buf, nil
+}
+
+func (a *indexEntries) WriteTo(w io.Writer) (total int64, err error) {
+	var buf [indexEntrySize]byte
+	var n int
+
+	for _, entry := range a.entries {
+		entry.AppendTo(buf[:])
+		n, err = w.Write(buf[:])
+		total += int64(n)
+		if err != nil {
+			return total, err
+		}
+	}
+
+	return total, nil
+}
+
 func (d *directIndex) Add(key []byte, blockType byte, minTime, maxTime int64, offset int64, size uint32) {
 	// Is this the first block being added?
 	if len(d.key) == 0 {
